See commit description
Implemented a way of authorising clients on the /send-data client.
The way I've done is probably not how it will be done in the end,
and I'm pretty sure it's is vulnerable to timing attacks. Just trying
to figure out Rust and Warp at this point.

Also ignore the SSL certificate/key pair from the previous commit,
I accidentally checked them in but I'm not using them anymore.